refactor: 레이아웃 관련 관련 컴포넌트 디렉토리 수정
